@@ -1,5 +1,5 @@
 [build-system]
-requires = ["poetry>=0.12"]
+requires = ["poetry>=1.7"]
 build-backend = "poetry.masonry.api"
 
 [tool.poetry]
@@ -11,47 +11,48 @@ authors = ["CTMS Reviewers <@mozilla-it/ctms-reviewers>"]
 
 [tool.poetry.dependencies]
 # These packages are mandatory and form the core of this package’s distribution.
-python = "~3.11.1"
-fastapi = "^0.84.0"
-starlette = "^0.19.1"
-requests = "^2.31.0"
-uvicorn = {extras = ["standard"], version = "^0.23.2"}
-pydantic = {extras = ["email"], version = "^1.10.12"}
-SQLAlchemy = ">1.3.23,<1.5"
-alembic = "^1.12.0"
-python-multipart = "^0.0.6"
+python = "^3.12.1"
+fastapi = "^0.108.0"
+starlette = "^0.32.0"
+requests = "^2.32.3"
+pydantic = {extras = ["email"], version = "^2.8.2"}
+uvicorn = {extras = ["standard"], version = "^0.30.3"}
+SQLAlchemy = "^2.0.31"
+alembic = "^1.13.2"
+python-multipart = "^0.0.9"
 python-jose = {extras = ["cryptography"], version = "^3.2.0"}
 passlib = {extras = ["argon2"], version = "^1.7.4"}
-python-dateutil = "^2.8.1"
+python-dateutil = "^2.9.0"
 dockerflow = "^2022.8.0"
-sentry-sdk = "^1.31.0"
-pysilverpop = "^0.2.6"
-lxml = "^4.9.3"
-prometheus-client = "0.17.1"
-structlog = "^23.1.0"
+sentry-sdk = "^2.11.0"
+lxml = "^5.2.2"
+prometheus-client = "0.20.0"
+structlog = "^24.4.0"
 colorama = "^0.4.6"
-psycopg2 = "^2.9.7"
-google-auth = "^2.23.0"
+psycopg2 = "^2.9.9"
+google-auth = "^2.32.0"
+pydantic-settings = "^2.3.4"
 
 
-[tool.poetry.dev-dependencies]
-pre-commit = "^3.4.0"
-sphinx = "^7.2.6"
-pytest = "^7.4.2"
-isort = {extras = ["pyproject"], version = "^5.12.0"}
-black = "^23.9.1"
-coverage = {extras = ["toml"], version = "^7.3"}
-mypy = "^1.5"
-types-python-dateutil = "^2.8.19"
-detect-secrets = "^1.4.0"
-bandit = "^1.7.5"
-SQLAlchemy-Utils = "^0.41.1"
-pylint = "^2.17.5"
-pylint-pytest = "^1.1.2"
-types-requests = "^2.31.0"
+[tool.poetry.group.dev.dependencies]
+pre-commit = "^3.8.0"
+sphinx = "^7.4.7"
+pytest = "^8.2.0"
+isort = {extras = ["pyproject"], version = "^5.13.2"}
+black = "^24.4.2"
+coverage = {extras = ["toml"], version = "^7.6"}
+mypy = "^1.11"
+types-python-dateutil = "^2.9.0"
+detect-secrets = "^1.5.0"
+bandit = "^1.7.9"
+SQLAlchemy-Utils = "^0.41.2"
+pylint = "^3.2.6"
+pylint-pytest = "^1.1.8"
+types-requests = "^2.32.0"
 factory-boy = "^3.3.0"
-pytest-factoryboy = "^2.5.1"
+pytest-factoryboy = "^2.7.0"
 backoff = "^2.2.1"
+httpx = "^0.27.0"
 
 [tool.pytest.ini_options]
 testpaths = [
@@ -118,7 +119,7 @@ profile = "black"
 
 [tool.mypy]
 # Sync Python version with Dockerfile
-python_version = "3.11"
+python_version = "3.12"
 # Look for errors in this file
 warn_unused_configs = true
 plugins = ["pydantic.mypy"]
