@@ -1,17 +1,14 @@
-from typing import Optional, cast
-from uuid import UUID as PythonUUID
+from uuid import UUID as UUID4
 
 from sqlalchemy import (
     JSON,
     TIMESTAMP,
     Boolean,
-    Column,
     Date,
     DateTime,
     ForeignKey,
     Index,
     Integer,
-    PrimaryKeyConstraint,
     String,
     Text,
     UniqueConstraint,
@@ -19,13 +16,17 @@ from sqlalchemy import (
 from sqlalchemy.dialects.postgresql import UUID
 from sqlalchemy.ext.declarative import declared_attr
 from sqlalchemy.ext.hybrid import Comparator, hybrid_property
-from sqlalchemy.orm import relationship
+from sqlalchemy.orm import DeclarativeBase, Mapped, mapped_column, relationship
 from sqlalchemy.sql.functions import func
 
-from .database import Base
+
+class Base(DeclarativeBase):
+    pass
 
 
-class CaseInsensitiveComparator(Comparator):  # pylint: disable=abstract-method
+class CaseInsensitiveComparator(
+    Comparator
+):  # pylint: disable=abstract-method,too-many-ancestors
     def __eq__(self, other):
         return func.lower(self.__clause_element__()) == func.lower(other)
 
@@ -33,19 +34,21 @@ class CaseInsensitiveComparator(Comparator):  # pylint: disable=abstract-method
 class TimestampMixin:
     @declared_attr
     def create_timestamp(cls):  # pylint: disable=no-self-argument
-        return Column(
-            TIMESTAMP(timezone=True), nullable=False, server_default=func.now()
+        return mapped_column(
+            TIMESTAMP(timezone=True),
+            nullable=False,
+            server_default=func.now(),  # pylint: disable=not-callable
         )
 
     @declared_attr
     def update_timestamp(cls):  # pylint: disable=no-self-argument
-        return Column(
+        return mapped_column(
             TIMESTAMP(timezone=True),
             nullable=False,
-            server_default=func.now(),
+            server_default=func.now(),  # pylint: disable=not-callable
             # server_onupdate would be nice to use here, but it's not supported
             # by Postgres
-            onupdate=func.now(),
+            onupdate=func.now(),  # pylint: disable=not-callable
         )
 
 
@@ -53,18 +56,18 @@ class Email(Base, TimestampMixin):
     __tablename__ = "emails"
     __mapper_args__ = {"eager_defaults": True}
 
-    email_id = Column(UUID(as_uuid=True), primary_key=True)
-    primary_email = Column(String(255), unique=True, nullable=False)
-    basket_token = Column(String(255), unique=True)
-    sfdc_id = Column(String(255), index=True)
-    first_name = Column(String(255))
-    last_name = Column(String(255))
-    mailing_country = Column(String(255))
-    email_format = Column(String(1))
-    email_lang = Column(String(5))
-    double_opt_in = Column(Boolean)
-    has_opted_out_of_email = Column(Boolean)
-    unsubscribe_reason = Column(Text)
+    email_id: Mapped[UUID4] = mapped_column(UUID(as_uuid=True), primary_key=True)
+    primary_email = mapped_column(String(255), unique=True, nullable=False)
+    basket_token = mapped_column(String(255), unique=True)
+    sfdc_id = mapped_column(String(255), index=True)
+    first_name = mapped_column(String(255))
+    last_name = mapped_column(String(255))
+    mailing_country = mapped_column(String(255))
+    email_format = mapped_column(String(1))
+    email_lang = mapped_column(String(5))
+    double_opt_in = mapped_column(Boolean)
+    has_opted_out_of_email = mapped_column(Boolean)
+    unsubscribe_reason = mapped_column(Text)
 
     newsletters = relationship(
         "Newsletter", back_populates="email", order_by="Newsletter.name"
@@ -76,15 +79,6 @@ class Email(Base, TimestampMixin):
     amo = relationship("AmoAccount", back_populates="email", uselist=False)
     mofo = relationship(
         "MozillaFoundationContact", back_populates="email", uselist=False
-    )
-    stripe_customer = relationship(
-        "StripeCustomer",
-        uselist=False,
-        back_populates="email",
-        overlaps="fxa",
-        primaryjoin="Email.email_id==FirefoxAccount.email_id",
-        secondaryjoin="remote(FirefoxAccount.fxa_id)==foreign(StripeCustomer.fxa_id)",
-        secondary="join(FirefoxAccount, StripeCustomer, FirefoxAccount.fxa_id == StripeCustomer.fxa_id)",
     )
 
     # Class Comparators
@@ -106,14 +100,16 @@ class Email(Base, TimestampMixin):
 class Newsletter(Base, TimestampMixin):
     __tablename__ = "newsletters"
 
-    id = Column(Integer, primary_key=True)
-    email_id = Column(UUID(as_uuid=True), ForeignKey(Email.email_id), nullable=False)
-    name = Column(String(255), nullable=False)
-    subscribed = Column(Boolean)
-    format = Column(String(1))
-    lang = Column(String(5))
-    source = Column(Text)
-    unsub_reason = Column(Text)
+    id = mapped_column(Integer, primary_key=True)
+    email_id: Mapped[UUID4] = mapped_column(
+        UUID(as_uuid=True), ForeignKey(Email.email_id), nullable=False
+    )
+    name = mapped_column(String(255), nullable=False)
+    subscribed = mapped_column(Boolean)
+    format = mapped_column(String(1))
+    lang = mapped_column(String(5))
+    source = mapped_column(Text)
+    unsub_reason = mapped_column(Text)
 
     email = relationship("Email", back_populates="newsletters", uselist=False)
 
@@ -123,13 +119,15 @@ class Newsletter(Base, TimestampMixin):
 class Waitlist(Base, TimestampMixin):
     __tablename__ = "waitlists"
 
-    id = Column(Integer, primary_key=True)
-    email_id = Column(UUID(as_uuid=True), ForeignKey(Email.email_id), nullable=False)
-    name = Column(String(255), nullable=False)
-    source = Column(Text)
-    subscribed = Column(Boolean, nullable=False, default=True)
-    unsub_reason = Column(Text)
-    fields = Column(JSON, nullable=False, server_default="'{}'::json")
+    id = mapped_column(Integer, primary_key=True)
+    email_id: Mapped[UUID4] = mapped_column(
+        UUID(as_uuid=True), ForeignKey(Email.email_id), nullable=False
+    )
+    name = mapped_column(String(255), nullable=False)
+    source = mapped_column(Text)
+    subscribed = mapped_column(Boolean, nullable=False, default=True)
+    unsub_reason = mapped_column(Text)
+    fields = mapped_column(JSON, nullable=False, server_default="'{}'::json")
 
     email = relationship("Email", back_populates="waitlists", uselist=False)
 
@@ -139,26 +137,18 @@ class Waitlist(Base, TimestampMixin):
 class FirefoxAccount(Base, TimestampMixin):
     __tablename__ = "fxa"
 
-    id = Column(Integer, primary_key=True)
-    fxa_id = Column(String(255), unique=True)
-    email_id = Column(
+    id = mapped_column(Integer, primary_key=True)
+    fxa_id = mapped_column(String(255), unique=True)
+    email_id = mapped_column(
         UUID(as_uuid=True), ForeignKey(Email.email_id), unique=True, nullable=False
     )
-    primary_email = Column(String(255), index=True)
-    created_date = Column(String(50))
-    lang = Column(String(255))
-    first_service = Column(String(50))
-    account_deleted = Column(Boolean)
+    primary_email = mapped_column(String(255), index=True)
+    created_date = mapped_column(String(50))
+    lang = mapped_column(String(255))
+    first_service = mapped_column(String(50))
+    account_deleted = mapped_column(Boolean)
 
-    email = relationship(
-        "Email", back_populates="fxa", overlaps="stripe_customer", uselist=False
-    )
-    stripe_customer = relationship(
-        "StripeCustomer",
-        uselist=False,
-        viewonly=True,
-        primaryjoin=("foreign(FirefoxAccount.fxa_id)==remote(StripeCustomer.fxa_id)"),
-    )
+    email = relationship("Email", back_populates="fxa", uselist=False)
 
     # Class Comparators
     @hybrid_property
@@ -178,40 +168,22 @@ class FirefoxAccount(Base, TimestampMixin):
 class AmoAccount(Base, TimestampMixin):
     __tablename__ = "amo"
 
-    id = Column(Integer, primary_key=True)
-    email_id = Column(
+    id = mapped_column(Integer, primary_key=True)
+    email_id = mapped_column(
         UUID(as_uuid=True), ForeignKey(Email.email_id), unique=True, nullable=False
     )
-    add_on_ids = Column(String(500))
-    display_name = Column(String(255))
-    email_opt_in = Column(Boolean)
-    language = Column(String(5))
-    last_login = Column(Date)
-    location = Column(String(255))
-    profile_url = Column(String(40))
-    user = Column(Boolean)
-    user_id = Column(String(40), index=True)
-    username = Column(String(100))
+    add_on_ids = mapped_column(String(500))
+    display_name = mapped_column(String(255))
+    email_opt_in = mapped_column(Boolean)
+    language = mapped_column(String(5))
+    last_login = mapped_column(Date)
+    location = mapped_column(String(255))
+    profile_url = mapped_column(String(40))
+    user = mapped_column(Boolean)
+    user_id = mapped_column(String(40), index=True)
+    username = mapped_column(String(100))
 
     email = relationship("Email", back_populates="amo", uselist=False)
-
-
-class AcousticField(Base):
-    __tablename__ = "acoustic_field"
-
-    tablename = Column(String, default="main")
-    field = Column(String)
-
-    __table_args__ = (
-        PrimaryKeyConstraint("tablename", "field", name="pk_tablename_field"),
-    )
-
-
-class AcousticNewsletterMapping(Base):
-    __tablename__ = "acoustic_newsletter_mapping"
-
-    source = Column(String, primary_key=True)
-    destination = Column(String)
 
 
 class ApiClient(Base, TimestampMixin):
@@ -219,263 +191,22 @@ class ApiClient(Base, TimestampMixin):
 
     __tablename__ = "api_client"
 
-    client_id = Column(String(255), primary_key=True)
-    email = Column(String(255), nullable=False)
-    enabled = Column(Boolean, default=True)
-    hashed_secret = Column(String, nullable=False)
+    client_id = mapped_column(String(255), primary_key=True)
+    email = mapped_column(String(255), nullable=False)
+    enabled = mapped_column(Boolean, default=True)
+    hashed_secret = mapped_column(String, nullable=False)
+    last_access = mapped_column(DateTime(timezone=True))
 
 
 class MozillaFoundationContact(Base, TimestampMixin):
     __tablename__ = "mofo"
 
-    id = Column(Integer, primary_key=True)
-    email_id = Column(
+    id = mapped_column(Integer, primary_key=True)
+    email_id = mapped_column(
         UUID(as_uuid=True), ForeignKey(Email.email_id), unique=True, nullable=False
     )
-    mofo_email_id = Column(String(255), unique=True)
-    mofo_contact_id = Column(String(255), index=True)
-    mofo_relevant = Column(Boolean)
+    mofo_email_id = mapped_column(String(255), unique=True)
+    mofo_contact_id = mapped_column(String(255), index=True)
+    mofo_relevant = mapped_column(Boolean)
 
     email = relationship("Email", back_populates="mofo", uselist=False)
-
-
-class PendingAcousticRecord(Base, TimestampMixin):
-    __tablename__ = "pending_acoustic"
-
-    id = Column(Integer, primary_key=True)
-    email_id = Column(
-        UUID(as_uuid=True), ForeignKey(Email.email_id), unique=True, nullable=False
-    )
-    retry = Column(Integer, nullable=False, default=0)
-    last_error = Column(Text, default="")
-
-    email = relationship("Email", uselist=False)
-
-
-class StripeBase(Base):
-    """Base class for Stripe objects."""
-
-    __abstract__ = True
-
-    def get_email_id(self) -> Optional[PythonUUID]:
-        """Return the email_id of the associated contact, if any."""
-        raise NotImplementedError()
-
-
-class StripeCustomer(StripeBase, TimestampMixin):
-    __tablename__ = "stripe_customer"
-
-    stripe_id = Column(String(255), nullable=False, primary_key=True)
-    fxa_id = Column(String(255), nullable=False, unique=True, index=True)
-    default_source_id = Column(String(255), nullable=True)
-    invoice_settings_default_payment_method_id = Column(String(255), nullable=True)
-
-    stripe_created = Column(DateTime(timezone=True), nullable=False)
-    deleted = Column(Boolean, nullable=False, default=False)
-
-    email = relationship(
-        "Email",
-        uselist=False,
-        back_populates="stripe_customer",
-        overlaps="email,fxa",
-        primaryjoin="remote(FirefoxAccount.fxa_id)==foreign(StripeCustomer.fxa_id)",
-        secondaryjoin="remote(Email.email_id)==foreign(FirefoxAccount.email_id)",
-        secondary="join(FirefoxAccount, StripeCustomer, FirefoxAccount.fxa_id == StripeCustomer.fxa_id)",
-    )
-    fxa = relationship(
-        "FirefoxAccount",
-        uselist=False,
-        viewonly=True,
-        primaryjoin=("remote(FirefoxAccount.fxa_id)==foreign(StripeCustomer.fxa_id)"),
-    )
-    invoices = relationship(
-        "StripeInvoice",
-        uselist=True,
-        viewonly=True,
-        primaryjoin=(
-            "foreign(StripeCustomer.stripe_id) =="
-            " remote(StripeInvoice.stripe_customer_id)"
-        ),
-    )
-    subscriptions = relationship(
-        "StripeSubscription",
-        back_populates="customer",
-        uselist=True,
-        primaryjoin=(
-            "foreign(StripeCustomer.stripe_id) =="
-            " remote(StripeSubscription.stripe_customer_id)"
-        ),
-    )
-
-    def get_email_id(self) -> Optional[PythonUUID]:
-        if self.fxa:
-            return cast(PythonUUID, self.fxa.email.email_id)
-        return None
-
-
-class StripePrice(StripeBase, TimestampMixin):
-    __tablename__ = "stripe_price"
-
-    stripe_id = Column(String(255), nullable=False, primary_key=True)
-    stripe_product_id = Column(String(255), nullable=False)
-
-    stripe_created = Column(DateTime(timezone=True), nullable=False)
-    active = Column(Boolean, nullable=False, default=True)
-    currency = Column(String(3), nullable=False)
-    recurring_interval = Column(String(5), nullable=True)
-    recurring_interval_count = Column(Integer, nullable=True)
-    unit_amount = Column(Integer, nullable=True)
-
-    invoice_line_items = relationship(
-        "StripeInvoiceLineItem", back_populates="price", uselist=True
-    )
-    subscription_items = relationship(
-        "StripeSubscriptionItem", back_populates="price", uselist=True
-    )
-
-    def get_email_id(self) -> None:
-        """Prices can be related to multiple Customers, so return None."""
-        return None
-
-
-class StripeInvoice(StripeBase, TimestampMixin):
-    __tablename__ = "stripe_invoice"
-
-    stripe_id = Column(String(255), nullable=False, primary_key=True)
-    stripe_customer_id = Column(String(255), nullable=False)
-    default_payment_method_id = Column(String(255), nullable=True, default=None)
-    default_source_id = Column(String(255), nullable=True, default=None)
-
-    stripe_created = Column(DateTime(timezone=True), nullable=False)
-    currency = Column(String(3), nullable=False)
-    total = Column(Integer, nullable=False)
-    status = Column(String(15), nullable=False)
-
-    customer = relationship(
-        "StripeCustomer",
-        uselist=False,
-        viewonly=True,
-        primaryjoin=(
-            " remote(StripeCustomer.stripe_id) =="
-            "foreign(StripeInvoice.stripe_customer_id)"
-        ),
-    )
-    line_items = relationship(
-        "StripeInvoiceLineItem", back_populates="invoice", uselist=True
-    )
-
-    def get_email_id(self) -> Optional[PythonUUID]:
-        if self.customer:
-            return cast(PythonUUID, self.customer.get_email_id())
-        return None
-
-
-class StripeInvoiceLineItem(StripeBase, TimestampMixin):
-    __tablename__ = "stripe_invoice_line_item"
-
-    stripe_id = Column(String(255), nullable=False, primary_key=True)
-    stripe_invoice_id = Column(
-        String(255),
-        ForeignKey("stripe_invoice.stripe_id"),
-        nullable=False,
-        index=True,
-    )
-    stripe_type = Column(String(14), nullable=False)
-    stripe_price_id = Column(
-        String(255),
-        ForeignKey(StripePrice.stripe_id),
-        nullable=False,
-        index=True,
-    )
-    stripe_invoice_item_id = Column(String(255), nullable=True)
-    stripe_subscription_id = Column(String(255), nullable=True)
-    stripe_subscription_item_id = Column(String(255), nullable=True)
-    amount = Column(Integer, nullable=False)
-    currency = Column(String(3), nullable=False)
-
-    invoice = relationship("StripeInvoice", back_populates="line_items", uselist=False)
-    price = relationship(
-        "StripePrice", back_populates="invoice_line_items", uselist=False
-    )
-    subscription = relationship(
-        "StripeSubscription",
-        uselist=False,
-        primaryjoin=(
-            " remote(StripeSubscription.stripe_id) =="
-            "foreign(StripeInvoiceLineItem.stripe_subscription_id)"
-        ),
-    )
-    subscription_item = relationship(
-        "StripeSubscriptionItem",
-        uselist=False,
-        primaryjoin=(
-            " remote(StripeSubscriptionItem.stripe_id) =="
-            "foreign(StripeInvoiceLineItem.stripe_subscription_item_id)"
-        ),
-    )
-
-    def get_email_id(self) -> Optional[PythonUUID]:
-        return cast(Optional[PythonUUID], self.invoice.get_email_id())
-
-
-class StripeSubscription(StripeBase, TimestampMixin):
-    __tablename__ = "stripe_subscription"
-
-    stripe_id = Column(String(255), nullable=False, primary_key=True)
-    stripe_customer_id = Column(String(255), nullable=False)
-    default_payment_method_id = Column(String(255), nullable=True)
-    default_source_id = Column(String(255), nullable=True)
-
-    stripe_created = Column(DateTime(timezone=True), nullable=False)
-    cancel_at_period_end = Column(Boolean, nullable=False)
-    canceled_at = Column(DateTime(timezone=True), nullable=True)
-    current_period_end = Column(DateTime(timezone=True), nullable=False)
-    current_period_start = Column(DateTime(timezone=True), nullable=False)
-    ended_at = Column(DateTime(timezone=True), nullable=True)
-    start_date = Column(DateTime(timezone=True), nullable=False)
-    status = Column(String(20), nullable=False)
-
-    customer = relationship(
-        "StripeCustomer",
-        uselist=False,
-        primaryjoin=(
-            " remote(StripeCustomer.stripe_id)=="
-            "foreign(StripeSubscription.stripe_customer_id)"
-        ),
-    )
-    subscription_items = relationship(
-        "StripeSubscriptionItem", back_populates="subscription", uselist=True
-    )
-
-    def get_email_id(self) -> Optional[PythonUUID]:
-        if self.customer:
-            return cast(PythonUUID, self.customer.get_email_id())
-        return None
-
-
-class StripeSubscriptionItem(StripeBase, TimestampMixin):
-    __tablename__ = "stripe_subscription_item"
-
-    stripe_id = Column(String(255), nullable=False, primary_key=True)
-    stripe_subscription_id = Column(
-        String(255),
-        ForeignKey(StripeSubscription.stripe_id),
-        nullable=False,
-    )
-    stripe_price_id = Column(
-        String(255),
-        ForeignKey(StripePrice.stripe_id),
-        nullable=False,
-    )
-
-    stripe_created = Column(DateTime(timezone=True), nullable=False)
-
-    subscription = relationship(
-        "StripeSubscription", back_populates="subscription_items", uselist=False
-    )
-    price = relationship(
-        "StripePrice", back_populates="subscription_items", uselist=False
-    )
-
-    def get_email_id(self) -> Optional[PythonUUID]:
-        return cast(Optional[PythonUUID], self.subscription.get_email_id())
