@@ -3,4 +3,4 @@ from pydantic import BaseModel
 
 class ComparableBase(BaseModel):
     def is_default(self):
-        return len(self.dict(exclude_defaults=True)) == 0
+        return len(self.model_dump(exclude_defaults=True)) == 0
