@@ -1,5 +1,5 @@
-from sqlalchemy import create_engine, engine
-from sqlalchemy.orm import declarative_base, scoped_session, sessionmaker
+from sqlalchemy import create_engine
+from sqlalchemy.orm import scoped_session, sessionmaker
 
 from .config import Settings
 
@@ -16,8 +16,6 @@ def engine_factory(settings):
 
 
 engine = engine_factory(Settings())
-SessionLocal = sessionmaker(autocommit=False, autoflush=False, bind=engine)
+SessionLocal = sessionmaker(autoflush=False, bind=engine)
 # Used for testing
 ScopedSessionLocal = scoped_session(SessionLocal)
-
-Base = declarative_base()
