@@ -5,6 +5,7 @@ Revises: 8f93a0e590f0
 Create Date: 2023-09-15 10:11:28.618091
 
 """
+
 # pylint: disable=no-member invalid-name
 # no-member is triggered by alembic.op, which has dynamically added functions
 # invalid-name is triggered by migration file names with a date prefix
@@ -21,7 +22,16 @@ depends_on = None
 
 
 def upgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
+    # Delete duplicate `pending_record` entries, keeping the latest one (highest `id`)
+    op.execute(
+        """
+    DELETE FROM pending_acoustic
+    WHERE id NOT IN
+        (SELECT MAX(id)
+         FROM pending_acoustic
+         GROUP BY email_id)
+    """
+    )
     op.create_unique_constraint("uix_pr_email_id", "pending_acoustic", ["email_id"])
     # ### end Alembic commands ###
 
