@@ -1,8 +1,8 @@
-"""adding an index to updatetimestamp in Email table
+"""Add column to track API use
 
-Revision ID: 832a2a588c88
-Revises: c57f9a6084eb
-Create Date: 2021-04-27 22:29:57.766070
+Revision ID: f2041a868dca
+Revises: cea70cf9d626
+Create Date: 2024-05-27 18:11:06.463499
 
 """
 
@@ -15,21 +15,22 @@ import sqlalchemy as sa
 from alembic import op
 
 # revision identifiers, used by Alembic.
-revision = "832a2a588c88"  # pragma: allowlist secret
-down_revision = "c57f9a6084eb"  # pragma: allowlist secret
+revision = "f2041a868dca"  # pragma: allowlist secret
+down_revision = "cea70cf9d626"  # pragma: allowlist secret
 branch_labels = None
 depends_on = None
 
 
 def upgrade():
     # ### commands auto generated by Alembic - please adjust! ###
-    op.create_index(
-        op.f("ix_emails_update_timestamp"), "emails", ["update_timestamp"], unique=False
+    op.add_column(
+        "api_client",
+        sa.Column("last_access", sa.DateTime(timezone=True), nullable=True),
     )
     # ### end Alembic commands ###
 
 
 def downgrade():
     # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_index(op.f("ix_emails_update_timestamp"), table_name="emails")
+    op.drop_column("api_client", "last_access")
     # ### end Alembic commands ###
