@@ -6,8 +6,10 @@ def test_delete_contact_by_primary_email_not_found(client):
     assert resp.status_code == 404
 
 
-def test_delete_contact_by_primary_email(client, maximal_contact):
-    primary_email = maximal_contact.email.primary_email
+def test_delete_contact_by_primary_email(client, dbsession, email_factory):
+    primary_email = email_factory().primary_email
+    dbsession.commit()
+
     resp = client.delete(f"/ctms/{primary_email}")
     assert resp.status_code == 200
     resp = client.delete(f"/ctms/{primary_email}")
@@ -15,8 +17,10 @@ def test_delete_contact_by_primary_email(client, maximal_contact):
 
 
 def test_delete_contact_by_primary_email_with_basket_token_unset(
-    client, most_minimal_contact
+    client, dbsession, email_factory
 ):
-    primary_email = most_minimal_contact.email.primary_email
-    resp = client.delete(f"/ctms/{primary_email}")
+    email = email_factory(basket_token=None)
+    dbsession.commit()
+
+    resp = client.delete(f"/ctms/{email.primary_email}")
     assert resp.status_code == 200
