@@ -1,10 +1,6 @@
 """Unit tests for GET /ctms/{email_id}"""
-from uuid import uuid4
 
 import pytest
-from structlog.testing import capture_logs
-
-from ctms.models import Email
 
 
 def test_get_ctms_for_minimal_contact(client, dbsession, email_factory):
@@ -67,6 +63,8 @@ def test_get_ctms_for_minimal_contact(client, dbsession, email_factory):
                 "source": newsletter.source,
                 "subscribed": newsletter.subscribed,
                 "unsub_reason": newsletter.unsub_reason,
+                "create_timestamp": newsletter.create_timestamp.isoformat(),
+                "update_timestamp": newsletter.update_timestamp.isoformat(),
             }
         ],
         "status": "ok",
@@ -133,6 +131,8 @@ def test_get_ctms_for_maximal_contact(client, maximal_contact):
                 "source": "https://www.mozilla.org/en-US/contribute/studentambassadors/",
                 "subscribed": False,
                 "unsub_reason": "Graduated, don't have time for FSA",
+                "create_timestamp": "2010-01-01T08:04:00+00:00",
+                "update_timestamp": "2020-01-28T14:50:00+00:00",
             },
             {
                 "format": "T",
@@ -141,6 +141,8 @@ def test_get_ctms_for_maximal_contact(client, maximal_contact):
                 "source": "https://commonvoice.mozilla.org/fr",
                 "subscribed": True,
                 "unsub_reason": None,
+                "create_timestamp": "2010-01-01T08:04:00+00:00",
+                "update_timestamp": "2020-01-28T14:50:00+00:00",
             },
             {
                 "format": "H",
@@ -149,6 +151,8 @@ def test_get_ctms_for_maximal_contact(client, maximal_contact):
                 "source": "https://www.mozilla.org/fr/firefox/accounts/",
                 "subscribed": False,
                 "unsub_reason": "done with this mailing list",
+                "create_timestamp": "2010-01-01T08:04:00+00:00",
+                "update_timestamp": "2020-01-28T14:50:00+00:00",
             },
             {
                 "format": "H",
@@ -157,6 +161,8 @@ def test_get_ctms_for_maximal_contact(client, maximal_contact):
                 "source": None,
                 "subscribed": True,
                 "unsub_reason": None,
+                "create_timestamp": "2010-01-01T08:04:00+00:00",
+                "update_timestamp": "2020-01-28T14:50:00+00:00",
             },
             {
                 "format": "H",
@@ -165,6 +171,8 @@ def test_get_ctms_for_maximal_contact(client, maximal_contact):
                 "source": None,
                 "subscribed": True,
                 "unsub_reason": None,
+                "create_timestamp": "2010-01-01T08:04:00+00:00",
+                "update_timestamp": "2020-01-28T14:50:00+00:00",
             },
             {
                 "format": "H",
@@ -173,6 +181,8 @@ def test_get_ctms_for_maximal_contact(client, maximal_contact):
                 "source": None,
                 "subscribed": True,
                 "unsub_reason": None,
+                "create_timestamp": "2010-01-01T08:04:00+00:00",
+                "update_timestamp": "2020-01-28T14:50:00+00:00",
             },
             {
                 "format": "H",
@@ -181,6 +191,8 @@ def test_get_ctms_for_maximal_contact(client, maximal_contact):
                 "source": None,
                 "subscribed": True,
                 "unsub_reason": None,
+                "create_timestamp": "2010-01-01T08:04:00+00:00",
+                "update_timestamp": "2020-01-28T14:50:00+00:00",
             },
         ],
         "status": "ok",
@@ -193,6 +205,8 @@ def test_get_ctms_for_maximal_contact(client, maximal_contact):
                 "source": "https://a-software.mozilla.org/",
                 "subscribed": True,
                 "unsub_reason": None,
+                "create_timestamp": "2010-01-01T08:04:00+00:00",
+                "update_timestamp": "2020-01-28T14:50:00+00:00",
             },
             {
                 "fields": {"geo": "cn"},
@@ -200,6 +214,8 @@ def test_get_ctms_for_maximal_contact(client, maximal_contact):
                 "source": None,
                 "subscribed": True,
                 "unsub_reason": None,
+                "create_timestamp": "2010-01-01T08:04:00+00:00",
+                "update_timestamp": "2020-01-28T14:50:00+00:00",
             },
             {
                 "fields": {"geo": "fr", "platform": "win64"},
@@ -207,6 +223,8 @@ def test_get_ctms_for_maximal_contact(client, maximal_contact):
                 "source": "https://super-product.mozilla.org/",
                 "subscribed": True,
                 "unsub_reason": None,
+                "create_timestamp": "2010-01-01T08:04:00+00:00",
+                "update_timestamp": "2020-01-28T14:50:00+00:00",
             },
             {
                 "fields": {"geo": "ca", "platform": "windows,android"},
@@ -214,6 +232,8 @@ def test_get_ctms_for_maximal_contact(client, maximal_contact):
                 "source": None,
                 "subscribed": True,
                 "unsub_reason": None,
+                "create_timestamp": "2010-01-01T08:04:00+00:00",
+                "update_timestamp": "2020-01-28T14:50:00+00:00",
             },
         ],
     }
@@ -277,6 +297,8 @@ def test_get_ctms_for_api_example(client, example_contact):
                 "source": None,
                 "subscribed": True,
                 "unsub_reason": None,
+                "create_timestamp": "2020-12-05T19:21:50.908000+00:00",
+                "update_timestamp": "2021-02-04T15:36:57.511000+00:00",
             },
             {
                 "format": "H",
@@ -285,6 +307,8 @@ def test_get_ctms_for_api_example(client, example_contact):
                 "source": None,
                 "subscribed": True,
                 "unsub_reason": None,
+                "create_timestamp": "2020-12-05T19:21:50.908000+00:00",
+                "update_timestamp": "2021-02-04T15:36:57.511000+00:00",
             },
         ],
         "status": "ok",
@@ -297,6 +321,8 @@ def test_get_ctms_for_api_example(client, example_contact):
                 "source": None,
                 "subscribed": True,
                 "unsub_reason": None,
+                "create_timestamp": "2020-12-05T19:21:50.908000+00:00",
+                "update_timestamp": "2021-02-04T15:36:57.511000+00:00",
             },
             {
                 "fields": {"geo": "fr"},
@@ -304,6 +330,8 @@ def test_get_ctms_for_api_example(client, example_contact):
                 "source": None,
                 "subscribed": True,
                 "unsub_reason": None,
+                "create_timestamp": "2020-12-05T19:21:50.908000+00:00",
+                "update_timestamp": "2021-02-04T15:36:57.511000+00:00",
             },
             {
                 "fields": {"geo": "fr", "platform": "ios,mac"},
@@ -311,6 +339,8 @@ def test_get_ctms_for_api_example(client, example_contact):
                 "source": None,
                 "subscribed": True,
                 "unsub_reason": None,
+                "create_timestamp": "2020-12-05T19:21:50.908000+00:00",
+                "update_timestamp": "2021-02-04T15:36:57.511000+00:00",
             },
         ],
     }
@@ -322,37 +352,6 @@ def test_get_ctms_not_found(client, dbsession):
     resp = client.get(f"/ctms/{email_id}")
     assert resp.status_code == 404
     assert resp.json() == {"detail": "Unknown email_id"}
-
-
-def test_get_ctms_not_traced(client, example_contact):
-    """Most CTMS contacts are not traced."""
-    email_id = example_contact.email.email_id
-    with capture_logs() as caplog:
-        resp = client.get(f"/ctms/{email_id}")
-    assert resp.status_code == 200
-    assert len(caplog) == 1
-    assert "trace" not in caplog[0]
-
-
-def test_get_ctms_with_tracing(client, dbsession):
-    """The log parameter trace is set when a traced email is requested."""
-    email_id = uuid4()
-    email = "test+trace-me-mozilla-123@example.com"
-    record = Email(
-        email_id=email_id,
-        primary_email=email,
-        double_opt_in=False,
-        email_format="T",
-        has_opted_out_of_email=False,
-    )
-    dbsession.add(record)
-    dbsession.commit()
-    with capture_logs() as caplog:
-        resp = client.get(f"/ctms/{email_id}")
-    assert resp.status_code == 200
-    assert len(caplog) == 1
-    assert caplog[0]["trace"] == email
-    assert "trace_json" not in caplog[0]
 
 
 @pytest.mark.parametrize("waitlist_name", ["vpn", "relay"])
