@@ -1,14 +1,13 @@
 import json
 from pathlib import Path
-from unittest.mock import Mock, patch
+from unittest.mock import Mock
 
 import pytest
 from sqlalchemy.exc import TimeoutError as SQATimeoutError
 from structlog.testing import capture_logs
 
 from ctms.app import app
-from ctms.config import Settings
-from ctms.dependencies import get_db, get_settings
+from ctms.dependencies import get_db
 
 
 @pytest.fixture
@@ -22,21 +21,6 @@ def mock_db():
     app.dependency_overrides[get_db] = mock_get_db
     yield mocked_db
     del app.dependency_overrides[get_db]
-
-
-@pytest.fixture
-def test_settings():
-    """Set settings for heartbeat tests."""
-
-    settings = {
-        "acoustic_retry_limit": 5,
-        "acoustic_batch_limit": 25,
-        "acoustic_loop_min_secs": 10,
-    }
-
-    app.dependency_overrides[get_settings] = lambda: Settings(**settings)
-    yield settings
-    del app.dependency_overrides[get_settings]
 
 
 def test_read_root(anon_client):
@@ -77,37 +61,13 @@ def test_crash_unauthorized(anon_client):
     assert resp.json() == {"detail": "Not authenticated"}
 
 
-def test_exposed_configuration(anon_client):
-    resp = anon_client.get("/acoustic_configuration")
-    exposed = resp.json()
-
-    assert "fxa_lang" in exposed["sync_fields"]["main"]
-    assert exposed["newsletter_mappings"]["mozilla-rally"] == "sub_rally"
-
-
-def test_read_heartbeat(anon_client, test_settings):
+def test_read_heartbeat(anon_client):
     """The platform calls /__heartbeat__ to check backing services."""
     with capture_logs() as cap_logs:
         resp = anon_client.get("/__heartbeat__")
     assert resp.status_code == 200
     data = resp.json()
-    expected = {
-        "database": {
-            "up": True,
-            "time_ms": data["database"]["time_ms"],
-            "acoustic": {
-                "success": True,
-                "backlog": 0,
-                "max_backlog": None,
-                "retry_backlog": 0,
-                "max_retry_backlog": None,
-                "retry_limit": 5,
-                "batch_limit": 25,
-                "loop_min_sec": 10,
-                "time_ms": data["database"]["acoustic"]["time_ms"],
-            },
-        }
-    }
+    expected = {"database": {"up": True, "time_ms": data["database"]["time_ms"]}}
     assert data == expected
     assert len(cap_logs) == 1
 
@@ -118,65 +78,8 @@ def test_read_heartbeat_no_db_fails(anon_client, mock_db):
     resp = anon_client.get("/__heartbeat__")
     assert resp.status_code == 503
     data = resp.json()
-    expected = {
-        "database": {
-            "up": False,
-            "time_ms": data["database"]["time_ms"],
-            "acoustic": {},
-        }
-    }
+    expected = {"database": {"up": False, "time_ms": data["database"]["time_ms"]}}
     assert data == expected
-
-
-def test_read_heartbeat_acoustic_fails(anon_client, test_settings):
-    """/__heartbeat__ returns 200 when measuring the acoustic backlog fails."""
-    with patch(
-        "ctms.acoustic_service.get_all_acoustic_records_count",
-        side_effect=SQATimeoutError(),
-    ):
-        resp = anon_client.get("/__heartbeat__")
-    assert resp.status_code == 503
-    data = resp.json()
-    expected = {
-        "database": {
-            "up": True,
-            "time_ms": data["database"]["time_ms"],
-            "acoustic": {
-                "success": False,
-                "backlog": None,
-                "max_backlog": None,
-                "retry_backlog": None,
-                "max_retry_backlog": None,
-                "retry_limit": 5,
-                "batch_limit": 25,
-                "loop_min_sec": 10,
-                "time_ms": data["database"]["acoustic"]["time_ms"],
-            },
-        }
-    }
-    assert data == expected
-
-
-@pytest.mark.parametrize("backlog, retry_backlog", ((51, 1), (1, 51)))
-def test_read_heartbeat_backlog_over_limit(
-    anon_client, test_settings, backlog, retry_backlog
-):
-    """/__heartbeat__ returns 503 when measuring the acoustic backlog fails."""
-    test_settings["acoustic_max_backlog"] = 50
-    test_settings["acoustic_max_retry_backlog"] = 50
-    with patch(
-        "ctms.acoustic_service.get_all_acoustic_records_count", return_value=backlog
-    ), patch(
-        "ctms.acoustic_service.get_all_acoustic_retries_count",
-        return_value=retry_backlog,
-    ):
-        resp = anon_client.get("/__heartbeat__")
-    assert resp.status_code == 503
-    data = resp.json()
-    assert data["database"]["acoustic"]["backlog"] == backlog
-    assert data["database"]["acoustic"]["max_backlog"] == 50
-    assert data["database"]["acoustic"]["retry_backlog"] == retry_backlog
-    assert data["database"]["acoustic"]["max_retry_backlog"] == 50
 
 
 def test_read_health(anon_client):
